Add full runfiles tree support
The runfiles tree of a fuzzing_binary is now packaged correctly for
OSS-Fuzz.

For a fuzz test foo in workspace bar, the runfiles are resolved relative
to the foo.runfiles/bar directory in the output .tar. A wrapper script
ensures that the real fuzz test binary is called with the correct
working directory.

The workspace_name of the binary label attribute in fuzzing_binary is
needed in the construction of the binary's runfiles tree and thus added
to the FuzzingBinaryInfo provider.